backend/requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
sqlalchemy==2.0.45
pydantic==2.12.5
email-validator==2.3.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
reportlab==4.4.5
